requirements.txt
svg.path==1.1
